HolderRecycler package refactoring
