+++ src/main/java/com/agileandroid/ui/adapters/holder/HolderRecycler.java
@@ -1,4 +1,4 @@
-package com.agileandroid.ui.adapters.recycler.holder;
+package com.agileandroid.ui.adapters.holder;
 
 /**
  * Created by oscar.urbina on 8/28/15.
