+++ src/main/java/com/agileandroid/ui/adapters/recycler/resolver/ViewResolver.java
@@ -4,7 +4,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
+import com.agileandroid.ui.adapters.holder.HolderRecycler;
 
 /**
  * Created by oscar.urbina on 8/28/15.
