+++ src/main/java/com/agileandroid/ui/adapters/view/interactor/Interactor.java
@@ -1,7 +1,7 @@
 package com.agileandroid.ui.adapters.view.interactor;
 
 import com.agileandroid.ui.adapters.AgileAdapterBuilder;
-import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
+import com.agileandroid.ui.adapters.holder.HolderRecycler;
 
 /**
  * Created by oscar.urbina on 8/28/15.
