+++ src/main/java/com/agileandroid/ui/adapters/view/populator/Populator.java
@@ -1,7 +1,7 @@
 package com.agileandroid.ui.adapters.view.populator;
 
 import com.agileandroid.ui.adapters.AgileAdapterBuilder;
-import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
+import com.agileandroid.ui.adapters.holder.HolderRecycler;
 
 /**
  * Created by oscar.urbina on 9/2/15.
