+++ src/main/java/com/agileandroid/ui/adapters/view/resolver/ViewResolver.java
@@ -4,7 +4,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 
 import com.agileandroid.ui.adapters.AgileAdapterBuilder;
-import com.agileandroid.ui.adapters.recycler.holder.HolderRecycler;
+import com.agileandroid.ui.adapters.holder.HolderRecycler;
 import com.agileandroid.ui.adapters.view.interactor.Interactor;
 import com.agileandroid.ui.adapters.view.populator.Populator;
